+++ languages/default.nix
@@ -2,4 +2,5 @@
 {
   rust = pkgs.callPackage ./rust { inherit base; };
   python = pkgs.callPackage ./python { inherit base; };
+  terraform = pkgs.callPackage ./terraform { inherit base; };
 }