+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Add mkTerraformComponent which is a provider-agnostic terraform module intended for deployment.
 - Rust functions can now pass in shellInputs and shellHooks.
 - Make src path invariant for Python and Rust packages. This makes it cacheable for everyone, irrespective of
   their path to the repositories. See